+++ tests/Servers/UnixTest.php
@@ -9,9 +9,9 @@ use Innmind\Server\Control\{
     Server\Processes,
     Server\Volumes,
 };
-use Innmind\TimeContinuum\Earth\Clock;
+use Innmind\TimeContinuum\Clock;
 use Innmind\TimeWarp\Halt\Usleep;
-use Innmind\Stream\Streams;
+use Innmind\IO\IO;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 use PHPUnit\Framework\Attributes\Group;
 
@@ -22,9 +22,9 @@ class UnixTest extends TestCase
     public function testInterface()
     {
         $this->assertInstanceOf(Server::class, Unix::of(
-            new Clock,
-            Streams::fromAmbientAuthority(),
-            new Usleep,
+            Clock::live(),
+            IO::fromAmbientAuthority(),
+            Usleep::new(),
         ));
     }
 
@@ -33,9 +33,9 @@ class UnixTest extends TestCase
     public function testProcesses()
     {
         $this->assertInstanceOf(Processes::class, Unix::of(
-            new Clock,
-            Streams::fromAmbientAuthority(),
-            new Usleep,
+            Clock::live(),
+            IO::fromAmbientAuthority(),
+            Usleep::new(),
         )->processes());
     }
 
@@ -44,9 +44,9 @@ class UnixTest extends TestCase
     public function testVolumes()
     {
         $this->assertInstanceOf(Volumes::class, Unix::of(
-            new Clock,
-            Streams::fromAmbientAuthority(),
-            new Usleep,
+            Clock::live(),
+            IO::fromAmbientAuthority(),
+            Usleep::new(),
         )->volumes());
     }
 }